face_recognition_project/requirements.txt
streamlit
numpy
pandas
opencv-python-headless
pillow
tensorflow-cpu  # Gunakan versi CPU agar ringan dan file tidak terlalu besar
# torch (jika pakai pytorch)
# torchvision (jika pakai pytorch)